+++ ReleaseNotes.md
@@ -5,6 +5,7 @@ This page highlights some changes in the SDK.
 Not all changes will be listed, but you can always [compare by version tags](https://github.com/AquaticInformatics/aquarius-sdk-net/compare/v17.2.21...v17.2.25) to see the full source code difference.
 
 ### 22.3.0
+- Updated the service models for the AQUARIUS Samples 2022.09 release.
 - Updated the service models for the AQUARIUS Time-Series 2022.3 release.
 - Deprecated ExtendedAttribute and Tag applicabilty boolean flag properties in favor of a new Applicability enum collection.
 - **Source code change required:** Removed deprecated "LocationTagListServiceRequest & Response" classes, use "TagListServiceRequest" instead.